Give every command its own module
create_cli.py goes from 1083 lines to 54: the group, ten imports, and an add_command loop. Each
command now lives in olinda/cli/commands/<verb>.py, decorator stack and body moved verbatim — the only
textual change is `@cli.command(` becoming `@click.command(`.

Registration through `add_command` rather than the decorator is what makes this safe. No command
module imports the group, so the import graph is a DAG that a plausible edit cannot make cyclic —
which matters because `fit` imports five of its siblings to call their callbacks directly. The loop
walks COMMAND_PANELS, so a command missing from the help config is now a KeyError at startup instead
of a silently extra panel.

commands/__init__.py states the rule the package depends on: nothing from olinda outside olinda/cli
may be imported at module scope, not even olinda.console. create_cli imports all ten modules to
register them, so one slip would put a 12-second import on every `olinda --help`.

Verified: help output byte-identical across all eleven screens at 100 columns; 55 tests pass;
importing the CLI in a clean interpreter still pulls in none of the nine extras-only modules; and
setuptools discovers all ten subpackages, so a built wheel is not missing any of this.
